+++ webpack.config.js
@@ -35,7 +35,11 @@ module.exports = {
       },
       {
         test: /\.scss$/,
-        use: ['style-loader', 'css-loader', 'sass-loader'],
+        use: [
+          'style-loader',
+          'css-loader',
+          { loader: 'sass-loader', options: { api: 'modern-compiler' } },
+        ],
       },
       {
         test: /\.js$/,